+++ src/main/kotlin/com/atlassian/performance/tools/jiraactions/api/action/DisableRichTextEditor.kt
@@ -1,6 +1,6 @@
 package com.atlassian.performance.tools.jiraactions.api.action
 
-import com.atlassian.performance.tools.jiraactions.api.DISABLE_RTE
+import com.atlassian.performance.tools.jiraactions.api.DISABLE_RICH_TEXT_EDITOR
 import com.atlassian.performance.tools.jiraactions.api.WebJira
 import com.atlassian.performance.tools.jiraactions.api.measure.ActionMeter
 
@@ -9,7 +9,7 @@ class DisableRichTextEditor(
     private val meter: ActionMeter
 ) : Action {
     override fun run() {
-        meter.measure(DISABLE_RTE) {
+        meter.measure(DISABLE_RICH_TEXT_EDITOR) {
             jira.configureRichTextEditor().disable()
         }
     }